Started porting on top of OpenSSL 1.1.x
OpenSSL 1.1.x makes a lot of types opaque, which are referenced
directly from estlib's code. This commit makes the compilable with
the current OpenSSL version but gives no guarantees if the code
actually works. Some old features which are no longer supported
like NETSCAPE-storage format or TLS-SRP are just disabled in the code.